fixed issue video
